Writing URL mechanism for IPC initialization
URL should be used now by client to initialize its own IPC client. The URL
contains 4 parts: user, key, host, port. `user` is the name of the client who
may want to identify itself to the server, and `key` will be used to
authenticate IPC messages. `host:port` specifies the server location.

After this change, we have to progress to change proxy code, using a inbound
query/info mechanism to get enough info for itself. This is not easy and still
in progress.